Apply patch from Marco Wahl.
Who says:

"I found a bug which might not have occurred in the past since the
monitors were smaller.  The bug shows up when the number of lines of the
match window exceeds the number of available indexes (which is 61).  I
think the patch (which is just a switch from 'greater than' to 'smaller
than') in the attachment is a way to go."